move file inside the package folder
